Preparations for x64 support: Handled ""warning C4312: 'reinterpret_cast': conversion from 'uint32_t' to 'const GLvoid *' of greater size"" - uncritical since we're only dealing with none-pointers like offets in here

+++ Renderer/RendererBackend/OpenGLES2Renderer/include/OpenGLES2Renderer/RenderTarget/Framebuffer.inl
@@ -49,7 +49,7 @@ namespace OpenGLES2Renderer
 	//[-------------------------------------------------------]
 	inline void* Framebuffer::getInternalResourceHandle() const
 	{
-		return reinterpret_cast<void*>(mOpenGLES2Framebuffer);
+		return reinterpret_cast<void*>(static_cast<uintptr_t>(mOpenGLES2Framebuffer));
 	}
