+++ Renderer/RendererBackend/OpenGLRenderer/include/OpenGLRenderer/RenderTarget/Framebuffer.inl
@@ -54,7 +54,7 @@ namespace OpenGLRenderer
 	//[-------------------------------------------------------]
 	inline void* Framebuffer::getInternalResourceHandle() const
 	{
-		return reinterpret_cast<void*>(mOpenGLFramebuffer);
+		return reinterpret_cast<void*>(static_cast<uintptr_t>(mOpenGLFramebuffer));
 	}
